+++ urbanstats/features/extract_data.py
@@ -1,4 +1,4 @@
-from census_blocks import load_raw_census
+from urbanstats.data.census_blocks import load_raw_census
 
 from .feature import features
 from .within_distance import minimum_distance_by_block
