+++ urbanstats/features/within_distance.py
@@ -5,7 +5,7 @@ import tqdm
 from permacache import permacache
 from scipy.spatial import cKDTree
 
-from census_blocks import load_raw_census
+from urbanstats.data.census_blocks import load_raw_census
 from urbanstats.geometry.ellipse import Ellipse
