+++ urbanstats/weather/to_blocks.py
@@ -1,7 +1,7 @@
 import numpy as np
 from permacache import permacache, stable_hash
 
-from census_blocks import load_raw_census
+from urbanstats.data.census_blocks import load_raw_census
 from urbanstats.weather.era5 import bounding_boxes
 from urbanstats.weather.stats import era5_statistics
 from urbanstats.weather.weather_statistic import compute_statistics
